Adjust hero and header text style
Target specific headers within divs and change the style based on the reference sheet. Also added proper color to the hero section. Not sure if it is best practice to target specific headers with selectors or if it's better to just add a class to the desired header. Will learn more later.